muzu.tv: makes playlists/jukebox work on 10.1
this is a horrid ugly bodge. all it does is insert a playlist item at the beginning of the playlist to play a tiny file consisting of the muzu logo in order that the first item in the playlist doesn't require to be resolved by the plugin.

this seems to give everything enough time to clean up and therefore not lock up as the playlist tried to call the plugin again to resolve the url before the plugin has finished.